+++ tests/testthat/test-assign.R
@@ -53,35 +53,29 @@ test_that("test assign - bin reps", {
   expect_equal(nrow(report(data, type = bin_representative)), 531)
 })
 
-test_that("test assign - quality", {
+test_that("test assign - sample_distance", {
   # Create a new empty strollur object named 'example_dataset'
   data <- strollur::new_dataset(dataset_name = "example_dataset")
 
-  # Read quality data into data.frame
-  quality_data <-
-    strollur::read_quality(quality = strollur_example("tiny.qual"))
-  fasta_data <-
-    strollur::read_fasta(fasta = strollur_example("tiny.fasta"))
+  shared_file <- strollur_example("final.opti_mcc.shared")
+  dist_file <- strollur_example("final.opti_mcc.jclass.0.03.column.dist")
+  reference <- strollur::new_reference(name = "jclass estimator distances",
+                                       vendor = "mothur_v1.48.6")
 
-  # Add FASTA sequence data
-  strollur::add(data = data, table = fasta_data, type = "sequence")
-  strollur::assign(data, table = quality_data, type = "quality")
+  df <- read_mothur_shared(shared_file)
+  xdev_assign_bins(data, table = df, bin_type = "otu")
 
-  quality_report <- strollur::report(data, type = "quality")
+  sample_dists <- readr::read_table(dist_file,
+                                    col_names = FALSE,
+                                    show_col_types = FALSE)
+  strollur::assign(data, table = sample_dists, type = "sample_distance")
 
-  names <- c(
-    "M00967:43:000000000-A3JHG:1:1101:18327:1699",
-    "M00967:43:000000000-A3JHG:1:1101:14069:1827",
-    "M00967:43:000000000-A3JHG:1:1101:18044:1900"
-  )
+  expect_equal(count(data, type = "sample"), 19)
 
-  expect_equal(quality_report$sequence_name, names)
+  dists <- xdev_get_sample_distances(data)
 
-  score_1_15 <- c(2, 29, 29, 32, 32, 33, 32, 33, 33, 37, 37, 37, 38, 38, 38)
-  score_2_15 <- c(18, 32, 32, 30, 32, 33, 33, 35, 33, 37, 37, 33, 36, 38, 38)
-  score_3_15 <- c(33, 32, 31, 33, 33, 33, 32, 33, 33, 37, 37, 37, 38, 38, 38)
+  expect_equal(nrow(dists), 171)
+  expect_equal(ncol(dists), 3)
+  expect_equal(round(dists[1:3, 3], 2), c(0.47, 0.53, 0.55))
 
-  expect_equal(quality_report$quality_score[[1]][1:15], score_1_15)
-  expect_equal(quality_report$quality_score[[2]][1:15], score_2_15)
-  expect_equal(quality_report$quality_score[[3]][1:15], score_3_15)
 })