+++ tests/testthat/test-read_mothur.R
@@ -36,7 +36,8 @@ test_that("test read_mothur", {
     design = strollur_example("mouse.time.design"),
     otu_list = strollur_example("final.opti_mcc.list.gz"),
     sample_distance = strollur_example(
-        "final.opti_mcc.jclass.0.03.column.dist"),
+      "final.opti_mcc.jclass.0.03.column.dist"
+    ),
     dataset_name = "miseq_sop"
   )
 
@@ -129,7 +130,8 @@ test_that("test read taxonomy files", {
     sample_tree = strollur_example("final.opti_mcc.jclass.ave.tre"),
     sequence_tree = strollur_example("final.phylip.tre.gz"),
     sample_distance = strollur_example(
-                        "final.opti_mcc.jclass.0.03.column.dist"),
+      "final.opti_mcc.jclass.0.03.column.dist"
+    ),
     dataset_name = "miseq_sop"
   )